added scan duration and savepointing
now, there is also scan duration as another iterating scheme and not
only the frequency hopping. So the same frequency range can be scanned
again until the duration time is over.
And there is now savepointing to restart, when it was aborted.
currently, only the duration time can be resumed. not yet the single
frequency hop.
and expansion and typo fix in readme