added node id, geo location
added a way to set a unique but harder to trace id for the node. based
on the computer mac address, but hashed to disguise it.
also added a first way to get an automated geo reference based on the ip
address. but only as an alternative, because we asume free http access
won't be the standard