Add capability to send the resulting data via e-mail.
In workunit.xml there are some additional settings now. The user can
specify if e-mails with results should be sent to a particular
destination where they could be collected eventually.

The Attribute 'send' in workunit.xml determines if mails should be sent.
In 'email_server', 'email_user' and 'email_pass' the user is able to
specify the corresponing data to access his or her SMTP account for
sending the e-mails to the destination defined in 'email_destination'.

This is a practical solution if there is no service like BOINC available
to collect the data directly.