Basic parsing and storing in databese.db
Note that at this stage, multiple tables are used although they are not
connected with each other.

Note also that there is currently no check if the data is already in the
database. It is intended to remove duplicates in the cleaning process
but this feature is not implemented yet.