Add -s option to db2json
With -s the db2json tool outputs only the last set of data per client id
hash.
This is useful if there is too much data in the database that cannot be
processed fast enough. The new option limits the outgoing data and
ensures a certain coverage over the clients.