Try to load 'id' and 'ppm' as strings if int fails
Because in node/bigwhoop.py there is an implicit conversion from some
integer values to strings somewhere regarding 'container', those values
are stored as strings instead of integers as it would be right.

So if this parser fails to load the values as integers, it will try to
load them as strings. If that fails also, a further exception will be
thrown.